+++ HPXMLOutputReport/resources/constants.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+class BS
+  # Cost multiplier types
+  Fixed = 'Fixed'
+  WallAreaAboveGradeConditioned = 'Wall Area Above-Grade Conditioned'
+  WallAreaAboveGradeExterior = 'Wall Area Above-Grade Exterior'
+  WallAreaBelowGrade = 'Wall Area Below-Grade'
+  FloorAreaConditioned = 'Floor Area Conditioned'
+  FloorAreaAttic = 'Floor Area Attic'
+  FloorAreaLighting = 'Floor Area Lighting'
+  RoofArea = 'Roof Area'
+  WindowArea = 'Window Area'
+  DoorArea = 'Door Area'
+  DuctUnconditionedSurfaceArea = 'Duct Unconditioned Surface Area'
+  SizeHeatingSystem = 'Size Heating System'
+  SizeSecondaryHeatingSystem = 'Size Secondary Heating System'
+  SizeHeatPumpBackup = 'Size Heat Pump Backup'
+  SizeCoolingSystem = 'Size Cooling System'
+  SizeWaterHeater = 'Size Water Heater'
+  FlowRateMechanicalVentilation = 'Flow Rate Mechanical Ventilation'
+  SlabPerimeterExposedConditioned = 'Slab Perimeter Exposed Conditioned'
+  RimJoistAreaAboveGradeExterior = 'Rim Joist Area Above-Grade Exterior'
+end
