+++ HPXMLOutputReport/resources/hpxml.rb
@@ -1,0 +1,122 @@
+# frozen_string_literal: true
+
+class HPXML
+  def self.get_cost_multiplier(hpxml, cost_mult_type)
+    cost_mult = 0.0
+    if cost_mult_type == 'Fixed'
+      cost_mult += 1.0
+    elsif cost_mult_type == 'Wall Area Above-Grade Conditioned'
+      hpxml.walls.each do |wall|
+        next unless wall.is_thermal_boundary
+
+        cost_mult += wall.area
+      end
+    elsif cost_mult_type == 'Wall Area Above-Grade Exterior'
+      hpxml.walls.each do |wall|
+        next unless wall.is_exterior
+
+        cost_mult += wall.area
+      end
+    elsif cost_mult_type == 'Wall Area Below-Grade'
+      hpxml.foundation_walls.each do |foundation_wall|
+        next unless foundation_wall.is_exterior
+
+        cost_mult += foundation_wall.area
+      end
+    elsif cost_mult_type == 'Floor Area Conditioned'
+      cost_mult += hpxml.building_construction.conditioned_floor_area
+    elsif cost_mult_type == 'Floor Area Attic'
+      hpxml.frame_floors.each do |frame_floor|
+        next unless frame_floor.is_thermal_boundary
+        next unless frame_floor.is_interior
+        next unless frame_floor.is_ceiling
+        next unless [HPXML::LocationAtticVented, HPXML::LocationAtticUnvented].include?(frame_floor.exterior_adjacent_to)
+
+        cost_mult += frame_floor.area
+      end
+    elsif cost_mult_type == 'Floor Area Lighting'
+      if hpxml.lighting.interior_usage_multiplier != 0
+        cost_mult += hpxml.building_construction.conditioned_floor_area
+      end
+      hpxml.slabs.each do |slab|
+        next unless [HPXML::LocationGarage].include?(slab.interior_adjacent_to)
+        next if hpxml.lighting.garage_usage_multiplier == 0
+
+        cost_mult += slab.area
+      end
+    elsif cost_mult_type == 'Roof Area'
+      hpxml.roofs.each do |roof|
+        cost_mult += roof.area
+      end
+    elsif cost_mult_type == 'Window Area'
+      hpxml.windows.each do |window|
+        cost_mult += window.area
+      end
+    elsif cost_mult_type == 'Door Area'
+      hpxml.doors.each do |door|
+        cost_mult += door.area
+      end
+    elsif cost_mult_type == 'Duct Unconditioned Surface Area'
+      hpxml.hvac_distributions.each do |hvac_distribution|
+        hvac_distribution.ducts.each do |duct|
+          next if [HPXML::LocationLivingSpace, HPXML::LocationBasementConditioned].include?(duct.duct_location)
+
+          cost_mult += duct.duct_surface_area
+        end
+      end
+    elsif cost_mult_type == 'Size Heating System'
+      hpxml.heating_systems.each do |heating_system|
+        next if heating_system.id != 'HeatingSystem'
+
+        cost_mult += UnitConversions.convert(heating_system.heating_capacity, 'btu/hr', 'kbtu/hr')
+      end
+
+      hpxml.heat_pumps.each do |heat_pump|
+        cost_mult += UnitConversions.convert(heat_pump.heating_capacity, 'btu/hr', 'kbtu/hr')
+      end
+    elsif cost_mult_type == 'Size Secondary Heating System'
+      hpxml.heating_systems.each do |heating_system|
+        next if heating_system.id != 'SecondHeatingSystem'
+
+        cost_mult += UnitConversions.convert(heating_system.heating_capacity, 'btu/hr', 'kbtu/hr')
+      end
+    elsif cost_mult_type == 'Size Heat Pump Backup'
+      hpxml.heat_pumps.each do |heat_pump|
+        cost_mult += UnitConversions.convert(heat_pump.backup_heating_capacity, 'btu/hr', 'kbtu/hr')
+      end
+    elsif cost_mult_type == 'Size Cooling System'
+      hpxml.cooling_systems.each do |cooling_system|
+        cost_mult += UnitConversions.convert(cooling_system.cooling_capacity, 'btu/hr', 'kbtu/hr')
+      end
+
+      hpxml.heat_pumps.each do |heat_pump|
+        cost_mult += UnitConversions.convert(heat_pump.cooling_capacity, 'btu/hr', 'kbtu/hr')
+      end
+    elsif cost_mult_type == 'Size Water Heater'
+      hpxml.water_heating_systems.each do |water_heating_system|
+        next if water_heating_system.tank_volume.nil?
+
+        cost_mult += water_heating_system.tank_volume
+      end
+    elsif cost_mult_type == 'Flow Rate Mechanical Ventilation'
+      hpxml.ventilation_fans.each do |ventilation_fan|
+        next unless ventilation_fan.used_for_whole_building_ventilation
+
+        if ventilation_fan.rated_flow_rate
+          cost_mult += ventilation_fan.rated_flow_rate
+        end
+      end
+    elsif cost_mult_type == 'Slab Perimeter Exposed Conditioned'
+      hpxml.slabs.each do |slab|
+        next unless slab.is_exterior_thermal_boundary
+
+        cost_mult += slab.exposed_perimeter
+      end
+    elsif cost_mult_type == 'Rim Joist Area Above-Grade Exterior'
+      hpxml.rim_joists.each do |rim_joist|
+        cost_mult += rim_joist.area
+      end
+    end
+    return cost_mult
+  end
+end
