+++ SimulationOutputReport/tests/output_report_test.rb
@@ -155,25 +155,6 @@ class SimulationOutputReportTest < MiniTest::Test
     'Hot Water: Dishwasher (gal)',
     'Hot Water: Fixtures (gal)',
     'Hot Water: Distribution Waste (gal)',
-    'Building Summary: Fixed (1)',
-    'Building Summary: Wall Area Above-Grade Conditioned (ft^2)',
-    'Building Summary: Wall Area Above-Grade Exterior (ft^2)',
-    'Building Summary: Wall Area Below-Grade (ft^2)',
-    'Building Summary: Floor Area Conditioned (ft^2)',
-    'Building Summary: Floor Area Attic (ft^2)',
-    'Building Summary: Floor Area Lighting (ft^2)',
-    'Building Summary: Roof Area (ft^2)',
-    'Building Summary: Window Area (ft^2)',
-    'Building Summary: Door Area (ft^2)',
-    'Building Summary: Duct Unconditioned Surface Area (ft^2)',
-    'Building Summary: Size Heating System (kBtu/h)',
-    'Building Summary: Size Secondary Heating System (kBtu/h)',
-    'Building Summary: Size Heat Pump Backup (kBtu/h)',
-    'Building Summary: Size Cooling System (kBtu/h)',
-    'Building Summary: Size Water Heater (gal)',
-    'Building Summary: Flow Rate Mechanical Ventilation (cfm)',
-    'Building Summary: Slab Perimeter Exposed Conditioned (ft)',
-    'Building Summary: Rim Joist Area Above-Grade Exterior (ft^2)',
   ]
 
   BaseHPXMLTimeseriesColsFuels = [
